[T1-54] [T1-60] [Home page] display user location and charger]

+++ lib/repositories/power_output/data_models/power_output_data_model.dart
@@ -6,9 +6,9 @@ part 'power_output_data_model.g.dart';
 @freezed
 class PowerOutputDataModel with _$PowerOutputDataModel {
   const factory PowerOutputDataModel({
-    @JsonKey(name: 'output_value') required int outputValue,
+    @JsonKey(name: 'output_value') required double outputValue,
     @JsonKey(name: 'charging_speed') required String chargingSpeed,
-    @JsonKey(name: 'voltage') required int voltage,
+    @JsonKey(name: 'voltage') required double voltage,
   }) = _PowerOutputDataModel;
 
   factory PowerOutputDataModel.fromJson(Map<String, dynamic> json) =>
